Proxy: register proxy services after resolve_networking
In 08d02f900e, we introduced code to automatically determine the
upstream address that the reverse proxy should use to contact its
upstream if the ProxyServerRole and ProxyServiceRole share a network
in common.

This code was not guraranteed to be safe because it ran in
resolve_networking on the ProxyServiceRole model.  However,
resolve_networking might not have run yet on the ProxyServerRole
model.

Introduce new ordering so that we collect a list of ProxyServices at
resolve_networking time, but hold off on actually registering those
services with the ProxyConfig until the point at which the
ProxyServerRole becomes async_ready.

* ProxyConfig: add add_proxied_model to register a ProxyServiceRole
  model as needing attention.

* ProxyServiceRole: move code out of resolve_networking to resolve_model which now exists.  At that point add the ProxyServiceRole to the ProxyConfig as needing attention.

* ProxySericeRole: Update register_proxy_map to take a config.

* ProxyServerRole: At async_ready time, go call register_proxy_map on
  all the proxied models to collect their services.