Certbot DNS: remove no-autorenew
We would prefer that certbot not autorenew our certificates because it
does no good. Either we will get called to install the new certificate
and we could renew then, or we will not, and it would be better to let
the certificate fall out of issue.  However, certbot's behavior on
no-autorenew is pathological: it treats the existing certificate as
unexpired, no matter how old it is, blocking manual attempts to renew
with the certonly command.

Therefore, allow autorenewals.